+++ examples/fetch_tabix/fetch_tabix.py
@@ -5,7 +5,7 @@ import vcfpy
 # Open input, add FILTER header, and open output file
 reader = vcfpy.Reader.from_path('input.vcf.gz')
 writer = vcfpy.Writer.from_path(
-    reader.header, reader.samples, '/dev/stdout')
+    '/dev/stdout', reader.header, reader.samples)
 
 # Fetch region 20:1,110,694-1,230,237.  Note that the coordinates
 # in the API call are zero-based and describe half-open intervals.