+++ tests/test_writer_write_header.py
@@ -45,7 +45,7 @@ def header_samples():
 def test_write_header(header_samples, tmpdir_factory):
     path = tmpdir_factory.mktemp('write_header').join('out.vcf')
     header, samples = header_samples
-    w = writer.Writer.from_path(header, samples, path)
+    w = writer.Writer.from_path(path, header, samples)
     w.close()
     RESULT = path.read()
     EXPECTED = MEDIUM_HEADER