+++ vcfpy/writer.py
@@ -6,6 +6,7 @@ Currently, only writing to plain-text files is supported
 
 from . import parser
 from . import record
+from . import bgzf
 
 __author__ = 'Manuel Holtgrewe <manuel.holtgrewe@bihealth.de>'
 
@@ -56,14 +57,25 @@ class Writer:
     """
 
     @classmethod
-    def from_stream(klass, stream, header, samples, path=None):
+    def from_stream(klass, stream, header, samples, path=None,
+                    use_bgzf=None):
         """Create new :py:class:`Writer` from file
+
+        Note that for getting bgzf support, you have to pass in a stream
+        opened in binary mode.  Further, you either have to provide a ``path``
+        ending in ``".gz"`` or set ``use_bgzf=True``.  Otherwise, you will
+        get the notorious "TypeError: 'str' does not support the buffer
+        interface".
 
         :param stream: ``file``-like object to write to
         :param header: VCF header to use
         :param samples: SamplesInfos to use
         :param path: optional string with path to store (for display only)
+        :param use_bgzf: indicator whether to write bgzf to ``stream``
+            if ``True``, prevent if ``False``, interpret ``path`` if ``None``
         """
+        if use_bgzf or (use_bgzf is None and path and path.endswith('.gz')):
+            stream = bgzf.BgzfWriter(fileobj=stream)
         return Writer(stream, header, samples, path)
 
     @classmethod
@@ -76,11 +88,12 @@ class Writer:
         :param samples: SamplesInfos to use
         """
         path = str(path)
+        use_bgzf = False  # we already interpret path
         if path.endswith('.gz'):
-            raise NotImplementedError('Writing to bgzf not supported')
+            f = bgzf.BgzfWriter(filename=path)
         else:
             f = open(path, 'wt')
-        return klass.from_stream(f, header, samples, path)
+        return klass.from_stream(f, header, samples, path, use_bgzf=use_bgzf)
 
     def __init__(self, stream, header, samples, path=None):
         #: stream (``file``-like object) to read from
@@ -97,7 +110,7 @@ class Writer:
     def _write_header(self):
         """Write out the header"""
         for line in self.header.lines:
-            print(line.serialize(), sep='', file=self.stream)
+            print(line.serialize(), file=self.stream)
         if self.samples.names:
             print('\t'.join(
                 list(parser.REQUIRE_SAMPLE_HEADER) + self.samples.names),