+++ HISTORY.rst
@@ -5,8 +5,8 @@ History
 HEAD
 ----
 
-* Renamed `Reader.jump_to` to `Reader.fetch`
-* Adding `header_without_lines` function
+* Renamed ``Reader.jump_to`` to ``Reader.fetch``
+* Adding ``header_without_lines`` function
 * Generally extending API to make it esier to use
 * Upgrading dependencies, enabling pyup-bot
 * Greatly extending documentation