+++ HISTORY.rst
@@ -5,7 +5,8 @@ History
 HEAD
 ----
 
-* Renamed ``from_file`` to ``from_stream``
+* Reordered parameters in ``from_stream`` and ``from_file`` (#18)
+* Renamed ``from_file`` to ``from_stream`` (#18)
 * Renamed ``Reader.jump_to`` to ``Reader.fetch``
 * Adding ``header_without_lines`` function
 * Generally extending API to make it esier to use